Move linux-specific functions
This change moves linux specific functions to _linux.go files.
This enables basic builds and tests on other platforms.
