+++ pkg/cgroups/cgroupid_other.go
@@ -1,0 +1,8 @@
+//go:build !linux
+// +build !linux
+
+package cgroups
+
+func getID(path string) uint64 {
+	panic("not implemented")
+}
